SL-18330: Refactor notation parsing to manage a lookahead char.
Instead of peeking ahead by a character and then (most of the time) rereading
it simply to discard it, go ahead and read the next character, passing it into
the various LLSDNotationParser._parse_mumble() functions. This eliminates most
LLSDNotationParser use of LLSDBaseParser._peek(), save for _get_re().

This eliminates the need for LLSDNotationParser._skip_then(), whose only job
was to discard the next character and return the specified value.

Break out LLSDNotationParser._parse_true() and _parse_false() in anticipation
of potential reimplementation without _get_re().