SL-18330: In XML formatter, avoid adding call stack depth.
Making LLSDXMLFormatter._elt() accept a lambda was a tricky way to minimize
source changes to existing _elt() calls in _ARRAY() and _MAP(). The trouble is
that that added function entries for each level of a deeply-nested LLSD
structure -- and with our users, we unfortunately do encounter deeply-nested
large inventories. Log observed RecursionError failures in AIS.

Explicitly write out the individual sequence of calls in _ARRAY(), _MAP() and
the top-level _write(), the only callers to pass lambdas to _elt().